--- app/src/main/java/com/mataku/scrobscrob/app/model/entity/NowPlaying.kt
@@ -1,9 +1,0 @@
-package com.mataku.scrobscrob.app.model.entity
-
-data class NowPlaying(
-        val artist: NowPlayingArtist,
-        val ignoredMessage: IgnoredMessage,
-        val album: NowPlayingAlbum,
-        //        val albumArtist: NowPlayingArtist,
-        val track: NowPlayingTrack
-)
