+++ app/src/main/java/com/mataku/scrobscrob/app/model/entity/NowPlayingApiResponse.kt
@@ -2,7 +2,14 @@ package com.mataku.scrobscrob.app.model.entity
 
 import com.google.gson.annotations.SerializedName
 
-data class NowPlayingApiResponse(
-        @SerializedName("nowplaying")
-        val nowPlaying: NowPlaying
-)
+class NowPlayingApiResponse(@SerializedName("nowplaying")
+                            val nowPlaying: NowPlaying) {
+
+    data class NowPlaying(
+            val artist: NowPlayingArtist,
+            val ignoredMessage: IgnoredMessage,
+            val album: NowPlayingAlbum,
+            //        val albumArtist: NowPlayingArtist,
+            val track: NowPlayingTrack
+    )
+}
